requirements.txt
chainlit
dapr-agents>=0.7.1
python-dotenv
couchbase